chore: update s3 key format
